+++ changelog.md
@@ -1,6 +1,16 @@
 ====================================================================
 CHANGELOG
 ====================================================================
+## 2.0.5
+* NEW: Support PirateForms.
+* IMPROVED: Header cover on single product.
+* IMPROVED: WooCommerce header title.
+* IMPROVED: Add hook for section parts.
+* FIXED: Title on single post & WooCommerce category.
+* FIXED: Section background parallax effect.
+* FIXED: Parallax issue.
+* FIXED: Back to top button z-index.
+
 ## 2.0.4
 * REMOVED: Remove debug test.